+++ src/parser/error.rs
@@ -1,6 +1,8 @@
 use crate::parser::Input;
+use crate::parser::tokens::{newlines, token};
 use nom::error::ErrorKind as NomErrorKind;
 use std::fmt;
+use std::io;
 
 #[derive(PartialEq, Debug, Clone)]
 pub struct ParseError {
@@ -9,71 +11,88 @@ pub struct ParseError {
     kind: ErrorKind,
 }
 
-macro_rules! error_type {
-    ( $name:ident, $kind:expr ) => {
-        pub fn $name(error: nom::Err<(Input, ParseError)>) -> nom::Err<(Input, ParseError)> {
-            match error {
-                nom::Err::Error((input, _)) => ParseError::new(input, $kind),
-                e @ nom::Err::Incomplete(_) => e,
-                e @ nom::Err::Failure((_, _)) => e,
-            }
-        }
-    };
-
-    ( $name:ident, $kind:expr; $param_name:ident : $param_type:ty ) => {
-        pub fn $name($param_name: $param_type) -> impl Fn(nom::Err<(Input<'_>, ParseError)>) -> nom::Err<(Input<'_>, ParseError)> {
-            move |error|
-            match error {
-                nom::Err::Error((input, _)) => ParseError::new(input, $kind),
-                e @ nom::Err::Incomplete(_) => e,
-                e @ nom::Err::Failure((_, _)) => e,
-            }
-        }
-    };
-
-    
-    ( [ $name:ident, $kind:expr ] $( $error:pat => $result:expr ),+ ) => {
-        #[allow(unreachable_patterns)]
-        pub fn $name(error: nom::Err<(Input, ParseError)>) -> nom::Err<(Input, ParseError)> {
-            match error {
-                $( $error => $result, )+
-                nom::Err::Error((input, _)) => ParseError::new(input, $kind),
-                e @ nom::Err::Incomplete(_) => e,
-                e @ nom::Err::Failure((_, _)) => e,
-            }
-        }
-    };
-
-    ( token [ $name:ident ] $( $reserved:literal => $kind:expr ),+ ) => {
-        #[allow(unreachable_patterns)]
-        pub fn $name(error: nom::Err<(Input, ParseError)>) -> nom::Err<(Input, ParseError)> {
-            match error {
-                nom::Err::Error((input, error)) =>
-                    match error.kind {
-                        $( ErrorKind::Reserved($reserved) => ParseError::new(input, $kind), )+
-                        _ => ParseError::new_with(input, error)
-                    }
-                e @ nom::Err::Incomplete(_) => e,
-                e @ nom::Err::Failure((_, _)) => e,
-            }
+impl ParseError {
+    fn is_unhandled(&self) -> bool {
+        match self.kind {
+            ErrorKind::Reserved(_)
+            | ErrorKind::Char
+            | ErrorKind::String
+            | ErrorKind::Number
+            | ErrorKind::Keyword
+            | ErrorKind::UnknownNomError
+            | ErrorKind::UnhandledError => true,
+            _ => false,
         }
     }
-}
 
-impl ParseError {
-    fn new(input: Input<'_>, kind: ErrorKind) -> nom::Err<(Input<'_>, Self)> {
-        nom::Err::Error((
-            input,
-            ParseError {
-                column: input.column(),
-                line: input.line(),
-                kind,
-            },
-        ))
+    fn new_from(input: Input<'_>, error: Self, kind: ErrorKind) -> nom::Err<(Input<'_>, Self)> {
+        nom::Err::Error((input, ParseError { kind, ..error }))
     }
 
     fn new_with(input: Input<'_>, error: Self) -> nom::Err<(Input<'_>, Self)> {
         nom::Err::Error((input, error))
+    }
+
+    pub fn verbose_from_source(&self, source: &str, writer: &mut impl io::Write) -> io::Result<()> {
+        writeln!(writer, "{}", self)?;
+        writeln!(writer)?;
+        match self.kind {
+            ErrorKind::TerminatingParen(line, column) => {
+                writeln!(writer, "    |")?;
+                writeln!(
+                    writer,
+                    "{:>3} | {}",
+                    self.line,
+                    source
+                        .lines()
+                        .nth(self.line - 1)
+                        .unwrap_or("")
+                )?;
+                writeln!(
+                    writer,
+                    "    | {:>1$} error found here",
+                    "^",
+                    self.column - 1
+                )?;
+                writeln!(writer)?;
+                writeln!(writer, "    |")?;
+                writeln!(
+                    writer,
+                    "{:>3} | {}",
+                    line,
+                    source
+                        .lines()
+                        .nth(line - 1)
+                        .unwrap_or("")
+                )?;
+                writeln!(
+                    writer,
+                    "    | {:>1$} opening parenthesis found here",
+                    "^",
+                    column - 1
+                )?;
+            }
+            _ => {
+                writeln!(writer, "    |")?;
+                writeln!(
+                    writer,
+                    "{:>3} | {}",
+                    self.line,
+                    source
+                        .lines()
+                        .nth(self.line - 1)
+                        .unwrap_or("")
+                )?;
+                writeln!(
+                    writer,
+                    "    | {:>1$} error found here",
+                    "^",
+                    self.column - 1
+                )?;
+            }
+        }
+        writeln!(writer)?;
+        Ok(())
     }
 }
 
@@ -99,32 +118,27 @@ impl fmt::Display for ParseError {
         let error = match self.kind {
             // Token Errors
             ErrorKind::Reserved(reserved) => format!("missing '{}' token", reserved),
-            ErrorKind::Char => "error parsing char literal".into(),
+            ErrorKind::Char => "error parsing character literal".into(),
             ErrorKind::String => "error parsing string literal".into(),
             ErrorKind::Number => "error parsing number literal".into(),
-            ErrorKind::Identifier => "error parsing identifier".into(),
+            ErrorKind::Keyword => "found keyword where an identifier was expected".into(),
 
             // Expr Errors
-            ErrorKind::Literal => "error parsing literal".into(),
-            ErrorKind::TerminatingParen => "missing closing parentheses".into(),
-            ErrorKind::FnArrow => "missing '->' between function parameters and function body".into(),
+            ErrorKind::InvalidCharacter => "encountered invalid character".into(),
+            ErrorKind::TerminatingParen(_, _) => "missing closing parenthesis".into(),
+            ErrorKind::FnArrow => {
+                "missing '->' between function parameters and function body".into()
+            }
             ErrorKind::MatchBar => "missing '|' before match arm".into(),
             ErrorKind::MatchArrow => "missing '->' between match pattern and match body".into(),
             ErrorKind::MatchTo => "missing 'to' between match head and body".into(),
             ErrorKind::Then => "missing 'then' or '?' after if condition".into(),
-            ErrorKind::Else => "missing 'else' or ':' in if expression".into(),
+            ErrorKind::Else => "missing 'else' in if expression".into(),
             ErrorKind::LetAssign => "missing '=' in let assignment".into(),
             ErrorKind::LetIn => "missing 'in' in let expression".into(),
             ErrorKind::DelayAssign => "missing '=' in delay assignment".into(),
             ErrorKind::DelayIn => "missing 'in' in delay expression".into(),
-
-            // Match Errors
-            ErrorKind::BasicMatch => "error parsing a basic match".into(),
-            ErrorKind::StringMatch => "error parsing a string value match".into(),
-            ErrorKind::CharMatch => "error parsing a char value match".into(),
-            ErrorKind::NumberMatch => "error parsing a number value match".into(),
-            ErrorKind::IdentifierMatch => "error parsing an identifier match".into(),
-            ErrorKind::TerminatingParenMatch => "missing closing parentheses in match".into(),
+            ErrorKind::EndOfExpr => "unexpected end of expr".into(),
 
             // Decl Errors
             ErrorKind::DeclAssign => "missing '=' in expression declaration".into(),
@@ -133,16 +147,84 @@ impl fmt::Display for ParseError {
             ErrorKind::TerminatingNewline => "missing newline (expected here)".into(),
             ErrorKind::Eof => "reached end of file before parsing was completed".into(),
             ErrorKind::UnknownNomError => "unknown error from parsing".into(),
+            ErrorKind::UnhandledError => "unhandled parsing error".into(),
         };
-        write!(f, "error[E{:04}]: {} (found on line {}, column {})",
+        write!(
+            f,
+            "error[E{:04}]: {} (found on line {}, column {})",
             u16::from(self.kind),
             error,
             self.line,
-            self.column)
+            self.column
+        )
     }
 }
 
 impl<'a> std::error::Error for ParseError {}
+
+macro_rules! error_type {
+    ( $name:ident, $kind:expr ) => {
+        pub fn $name(error: nom::Err<(Input, ParseError)>) -> nom::Err<(Input, ParseError)> {
+            match error {
+                nom::Err::Error((input, error)) if error.is_unhandled() => ParseError::new_from(input, error, $kind),
+                e => e
+            }
+        }
+    };
+
+    ( $name:ident, $kind:expr; $param_name:ident : $param_type:ty ) => {
+        pub fn $name($param_name: $param_type) -> impl Fn(nom::Err<(Input<'_>, ParseError)>) -> nom::Err<(Input<'_>, ParseError)> {
+            move |error|
+            match error {
+                nom::Err::Error((input, error)) if error.is_unhandled() => ParseError::new_from(input, error, $kind),
+                e => e
+            }
+        }
+    };
+
+
+    ( [ $name:ident ] $( $error:pat $( if $cond:expr )? => $result:expr ),+ ) => {
+        #[allow(unreachable_patterns)]
+        pub fn $name(error: nom::Err<(Input, ParseError)>) -> nom::Err<(Input, ParseError)> {
+            match error {
+                $( $error $( if $cond:expr )? => $result, )+
+                nom::Err::Error((input, error)) if error.is_unhandled() => ParseError::new_from(input, error, ErrorKind::UnhandledError),
+                e => e
+            }
+        }
+    };
+
+    ( token [ $name:ident ] $( $reserved:literal => $kind:expr ),+ ) => {
+        #[allow(unreachable_patterns)]
+        pub fn $name(error: nom::Err<(Input, ParseError)>) -> nom::Err<(Input, ParseError)> {
+            match error {
+                nom::Err::Error((input, error)) if error.is_unhandled() =>
+                    match error.kind {
+                        $( ErrorKind::Reserved($reserved) => ParseError::new_from(input, error, $kind), )+
+                        _ => ParseError::new_with(input, error)
+                    },
+                e => e
+            }
+        }
+    };
+
+    ( starts_with [ $name:ident $( , $default_kind:expr)? ] $( $pattern:expr => $kind:expr ),+ ) => {
+        #[allow(unreachable_patterns)]
+        pub fn $name(error: nom::Err<(Input, ParseError)>) -> nom::Err<(Input, ParseError)> {
+            match error {
+                nom::Err::Error((input, error))
+                    if input.to_str().is_empty() || token(newlines(true))(input).is_ok() =>
+                        ParseError::new_from(input, error, ErrorKind::EndOfExpr),
+                $( nom::Err::Error((input, error))
+                    if error.is_unhandled() && input.to_str().starts_with($pattern) =>
+                        ParseError::new_from(input, error, $kind), )+
+                $( nom::Err::Error((input, error)) if error.is_unhandled() =>
+                    ParseError::new_from(input, error, $default_kind), )?
+                e => e
+            }
+        }
+    }
+}
 
 // Error handlers
 
@@ -151,25 +233,17 @@ error_type!(reserved_error, ErrorKind::Reserved(token); token: &'static str);
 error_type!(char_error, ErrorKind::Char);
 error_type!(string_error, ErrorKind::String);
 error_type!(number_error, ErrorKind::Number);
-error_type!(ident_error, ErrorKind::Identifier);
+error_type!(ident_error, ErrorKind::Keyword);
 
 // Expr Errors
 error_type! {
-    [literal_error, ErrorKind::Literal]
-    nom::Err::Error((input, _)) =>
-        if input.to_str().starts_with('"') {
-            ParseError::new(input, ErrorKind::String)
-        } else if input.to_str().starts_with('\'') {
-            ParseError::new(input, ErrorKind::Char)
-        } else if input.to_str().starts_with(|c: char| c.is_digit(10)) {
-            ParseError::new(input, ErrorKind::Number)
-        } else if input.to_str().starts_with(char::is_alphabetic) {
-            ParseError::new(input, ErrorKind::Identifier)
-        } else {
-            ParseError::new(input, ErrorKind::Literal)
-        }
+    starts_with [literal_error, ErrorKind::InvalidCharacter]
+    '"' => ErrorKind::String,
+    '\'' => ErrorKind::Char,
+    |c: char| c.is_digit(10) => ErrorKind::Number,
+    char::is_alphabetic => ErrorKind::Keyword
 }
-error_type!(terminating_paren_error, ErrorKind::TerminatingParen);
+error_type!(terminating_paren_error, ErrorKind::TerminatingParen(open_paren_loc.0, open_paren_loc.1); open_paren_loc: (usize, usize));
 error_type! {
     token [fn_expr_error]
     "->" => ErrorKind::FnArrow
@@ -205,23 +279,15 @@ error_type! {
 }
 
 // Match Errors
-error_type!(ident_match_error, ErrorKind::IdentifierMatch);
+error_type!(ident_match_error, ErrorKind::Keyword);
 error_type! {
-    [basic_match_error, ErrorKind::BasicMatch]
-    nom::Err::Error((input, _)) =>
-        if input.to_str().starts_with('"') {
-            ParseError::new(input, ErrorKind::StringMatch)
-        } else if input.to_str().starts_with('\'') {
-            ParseError::new(input, ErrorKind::CharMatch)
-        } else if input.to_str().starts_with(|c: char| c.is_digit(10)) {
-            ParseError::new(input, ErrorKind::NumberMatch)
-        } else if input.to_str().starts_with(char::is_alphabetic) {
-            ParseError::new(input, ErrorKind::IdentifierMatch)
-        } else {
-            ParseError::new(input, ErrorKind::BasicMatch)
-        }
+    starts_with [basic_match_error, ErrorKind::InvalidCharacter]
+    '"' => ErrorKind::String,
+    '\'' => ErrorKind::Char,
+    |c: char| c.is_digit(10) => ErrorKind::Number,
+    char::is_alphabetic => ErrorKind::Keyword
 }
-error_type!(grouping_match_error, ErrorKind::TerminatingParenMatch);
+error_type!(grouping_match_error, ErrorKind::TerminatingParen(open_paren_loc.0, open_paren_loc.1); open_paren_loc: (usize, usize));
 
 // Decl Errors
 error_type! {
@@ -239,11 +305,11 @@ enum ErrorKind {
     Char,
     String,
     Number,
-    Identifier,
+    Keyword,
+    InvalidCharacter,
 
     // Expr Errors
-    Literal,
-    TerminatingParen,
+    TerminatingParen(usize, usize),
     FnArrow,
     MatchBar,
     MatchArrow,
@@ -254,14 +320,7 @@ enum ErrorKind {
     LetIn,
     DelayAssign,
     DelayIn,
-
-    // Match Errors
-    BasicMatch,
-    StringMatch,
-    CharMatch,
-    NumberMatch,
-    IdentifierMatch,
-    TerminatingParenMatch,
+    EndOfExpr,
 
     // Decl Errors
     DeclAssign,
@@ -270,13 +329,14 @@ enum ErrorKind {
     TerminatingNewline,
     Eof,
     UnknownNomError,
+    UnhandledError,
 }
 
 impl From<NomErrorKind> for ErrorKind {
     fn from(error: NomErrorKind) -> Self {
         match error {
             NomErrorKind::Eof => ErrorKind::Eof,
-            _ => ErrorKind::UnknownNomError
+            _ => ErrorKind::UnknownNomError,
         }
     }
 }
@@ -295,9 +355,9 @@ impl From<&ErrorKind> for u16 {
             ErrorKind::Char => 2,
             ErrorKind::String => 3,
             ErrorKind::Number => 4,
-            ErrorKind::Identifier => 5,
-            ErrorKind::Literal => 6,
-            ErrorKind::TerminatingParen => 7,
+            ErrorKind::Keyword => 5,
+            ErrorKind::InvalidCharacter => 6,
+            ErrorKind::TerminatingParen(_, _) => 7,
             ErrorKind::FnArrow => 8,
             ErrorKind::MatchBar => 9,
             ErrorKind::MatchArrow => 10,
@@ -308,16 +368,12 @@ impl From<&ErrorKind> for u16 {
             ErrorKind::LetIn => 15,
             ErrorKind::DelayAssign => 16,
             ErrorKind::DelayIn => 17,
-            ErrorKind::BasicMatch => 18,
-            ErrorKind::StringMatch => 19,
-            ErrorKind::CharMatch => 20,
-            ErrorKind::NumberMatch => 21,
-            ErrorKind::IdentifierMatch => 22,
-            ErrorKind::TerminatingParenMatch => 23,
-            ErrorKind::DeclAssign => 24,
-            ErrorKind::TerminatingNewline => 25,
-            ErrorKind::Eof => 26,
-            ErrorKind::UnknownNomError => 27,
+            ErrorKind::DeclAssign => 18,
+            ErrorKind::TerminatingNewline => 19,
+            ErrorKind::Eof => 20,
+            ErrorKind::UnknownNomError => 21,
+            ErrorKind::UnhandledError => 22,
+            ErrorKind::EndOfExpr => 23
         }
     }
 }