+++ src/repl.rs
@@ -1,13 +1,16 @@
 use nom::combinator::all_consuming;
 use std::io::{self, Write};
 use std::rc::Rc;
-use tego_lang::ast::Decl;
-use tego_lang::interpreter;
-use tego_lang::parser;
+use tego::ast::Decl;
+use tego::interpreter;
+use tego::parser;
 
-pub fn run() {
-    println!("Welcome to");
-    println!(
+pub fn run() -> io::Result<()> {
+    let mut stdout = io::BufWriter::new(io::stdout());
+
+    writeln!(stdout, "Welcome to")?;
+    writeln!(
+        stdout,
         "
    /\\
   //\\\\
@@ -16,20 +19,25 @@ pub fn run() {
    ||  ||_  \\_|| \\\\_//
             \\_||       
 "
-    );
-    println!("Type ':q' or ':quit' to exit\n");
-    repl_loop(Some(interpreter::new_env()), vec![]);
+    )?;
+    writeln!(stdout, "Type ':q' or ':quit' to exit\n")?;
+    stdout.flush()?;
+    repl_loop(Some(interpreter::new_env()), vec![], stdout)
 }
 
-fn repl_loop(env: Option<interpreter::WrappedEnv>, mut decls: Vec<Decl>) {
-    print!(">> ");
-    io::stdout().flush().unwrap();
+fn repl_loop(
+    env: Option<interpreter::WrappedEnv>,
+    mut decls: Vec<Decl>,
+    mut stdout: io::BufWriter<io::Stdout>,
+) -> io::Result<()> {
+    write!(stdout, ">> ")?;
+    stdout.flush()?;
     let mut code = String::new();
     io::stdin().read_line(&mut code).unwrap();
     let code = code.trim();
 
     if code == ":quit" || code == ":q" {
-        return;
+        return Ok(());
     }
     let (env, decls) = if let Ok((_, d)) = parser::decl(code.into()) {
         decls.push(d);
@@ -39,18 +47,21 @@ fn repl_loop(env: Option<interpreter::WrappedEnv>, mut decls: Vec<Decl>) {
             Ok((_, e)) => {
                 let env = env.unwrap_or_else(|| interpreter::env_from_decls(&decls));
                 let result = interpreter::eval_expr(e, &Rc::clone(&env));
-                println!("{} : {}", result, result.type_());
+                writeln!(stdout, "{} : {}", result, result.type_())?;
                 (Some(env), decls)
             }
             Err(error) => {
                 match error {
-                    error @ nom::Err::Incomplete(_) => println!("{:?}", error),
-                    nom::Err::Failure((_, error)) => println!("{}", error),
-                    nom::Err::Error((_, error)) => println!("{}", error)
+                    error @ nom::Err::Incomplete(_) => writeln!(stdout, "{:?}", error)?,
+                    nom::Err::Failure((_, error)) => {
+                        error.verbose_from_source(code, &mut stdout)?
+                    }
+                    nom::Err::Error((_, error)) => error.verbose_from_source(code, &mut stdout)?,
                 }
                 (env, decls)
             }
         }
     };
-    repl_loop(env, decls);
+    stdout.flush()?;
+    repl_loop(env, decls, stdout)
 }